Add suggestion service and events.
This requires fake server 2.2.0 for fakewatch to work correctly.
This requires zanata-server version 3.7+ to work correctly in production.

This includes a stub implementation if the suggestion controller
that logs suggestions to the console.

Added an event for searching based on a Phrase, and for searching
based on text for when that is implemented.