Prevent suggestion search from dropping a search.
The search service could get into a state where there is no pending
search timer running, but the handle for the canceled timer is still
present. This caused the code to skip setting a timer because the
non-null timer was supposed to indicate that a timer was running.

This change just makes sure the handle is set to null every time the
timer is canceled.