Fix style to comply with initial eslint configuration.
So far this is just the default configuration with a few minor
customizations to match our current style.

Note: there are still several warnings generated because of logging
to console. A custom rule is needed to allow error logging but remove
other logging. I have left the rule active at warning level until a
custom rule can replace it.