Add typecheck task using flow.
Configured to check all javascript files in the app.

I do not know why, but this fails with an error unless there is
a .flowconfig under the project root, and one under the app
directory.

Some files directly under the app may be missed by the default app
path, this needs checking as some files may be missed by linting
and other tasks.