refactor: change documents dropdown to use React
DropdownService is handling 2 different types of dropdown now, in order
to ensure that only one is open. That service can be simplified significantly
when the last Angular dropdown is removed.

I chose a top-level property to determine whether dropdowns are open,
represented right now by DropdownService.getOpenDropdown(), to avoid storing
a boolean in each Dropdown state. Having multiple booleans would make it possible
to open two dropdowns at once, which is not the desired behaviour here (we can
easily change them to work independently if that is required, no need to complicate
the code with it now).