refactor: replace navigation header with React component
The directive for this component is starting to take the role of a store,
but still delegating most of its state management to Angular controllers
and services. Note that this single directive takes over from all the directives
of the components in the subtree - this pattern should continue until a single
store oversees the whole app.
Events are not yet handled with a dispatcher, so dropdown toggle and a few
other callbacks are passed into the component tree. These can be removed
when a dispatcher is added.