chore: inherit most .eslintrc rules from a shared file
There are a few differences in linting needed for the ES5/Angular code and the ES6/JSX/React code, but
most requirements are the same, so common.eslintrc was created to hold the shared parts, and inherited
in the others.
In the future, these config files could live in their own repository and be shared between zanata-spa
and other Zanata javascript repositories.