refactor: use module import for lodash
The npm lodash module still exports lodash as _ to the global scope.
For now that allows removal of the copy of lodash being packaged from
bower. Ultimately when no global lodash is wanted, an alternative
packaging should be found that does not pollute the global scope.
It appears as though this may be possible by just importing sub-modules
such as: import { pick } from 'lodash/object'