fix(dependency): lock angular to an unambiguous version
The required version for some angular plugins was incompatible with
the configured version in bower.json. When running the build, the
developer would be asked to choose from the 2 versions and the build
would fail if they did not select the rc.3 version.
The resolution is to specify the rc.3 version in bower.json. I chose
not to try updating all the angular plugins to the released version
since I expect angular to be removed from the build in the near
future, so any effort in updating it would be wasted.