refactor: replace diff switch with React component
Includes some changes to the webpack bundling:
 - change the directives in charge of React component trees to export modules
 - create a file that imports all the directives that build React component
   trees. This is the new entry point for webpack.
Trying to use a separate webpack build for each directive fails because there
are 2 instances of React, which causes problems. Trying a webpack build with
multiple entry points is more complexity than I want to deal with for this
temporary case. Using imports to combine both to a file gets around all these
problems.