refactor(IconButton): rename and fully test both IconButton components
The names were potentially confusing and did not really suggest that
IconButton would be an extension of PlainIconButton, so they were renamed:
 - change IconButton to IconButtonToggle
 - change PlainIconButton to IconButton
Tests were expanded to properly cover both cases, and the active property
is now limited to IconButtonToggle since showing that state is its main
differentiator from the base IconButton. Also ensured that clicks do not
trigger an event when the 'disabled' property is true, since that would
be surprising behaviour.