refactor: convert suggested translation detail to react
This adds the react-intl library, but does not set it up at the level
of the whole app yet. It should be configured at the top of the app
when react is used at the top of the component tree.
The react-intl library can take over all the functionality handled by
angular-gettext at the moment.