refactor: use redux for suggestion panel state management
Still leaves two Angular services that do the actual network requests for
suggestions. These can be swapped out at any time.
Also leaves the resizer component handled by Angular because it manipulates
a panel that is not under control of Redux yet, so it will be far easier to
migrate when the resizing state can be handled by redux.