WIP: connect paging and filtering
This hit a problem with the current reducer approach. The way the
reducers and state are structured, some of the reducers would need
access to the state handled by other reducers to derive values
without duplication in state.
I solved it by making getState() available in all actions with a
middleware. Other options are to rearrange the state or to take
a different approach to reducers - basically making a custom
combineReducers() that can pass extra state to the reducer in
addition to the state it is updating.