fix(focus phrase): focus correct field when a phrase is selected
Focuses the previously selected plural index (or the first plural form
if there was no previous selection) whenever a phrase changes from
not selected to selected. A click directly on the plural translation
will set it to the selected plural.
The React autoFocus property could not be used here because it would
steal focus from other components
The React autoFocus property could not be used here because it would
steal focus from other components.