refactor(gulpfile): prevent linting errors in gulpfile (quick)
Fixed a few minor errors, but just disabled most of the linting that
was causing errors since the gulpfile will be decommissioned soon.
I think it is better to leave it largely unchanged so that it is
easier to investigate how it reached its current state, even though
the style is a little inconsistent with the other javascript.
If it hangs around after Angular is removed, I will make it fully
style-compliant.