fix: bug caused by using && instead of ternary
This type of case is why I do not like using structures like this:

    const foo = bar && bar.baz

instead of structures like this:

    const foo = bar ? bar.baz : undefined

The former is a little shorter, but it is too easy to accidentally
break something by returning a boolean when a non-boolean or undefined
is expected.

Essentially, it is too easy from the habit of making the above substitution
to accidentally convert this:

    const elem = shouldHave ? <div>Yeah!</div> : undefined

to this:

    const elem = shouldHave && <div>Yeah!</div>

which is a big problem when elem must be some jsx or undefined, because
in this case its value will be either some jsx or boolean false.